+++ src/main.js
@@ -1,10 +1,15 @@
+import { getPoints } from './mock/point.js';
 import TripPointModel from './model/trip-point-model.js';
 import PointPresenter from './presenter/point-presenter.js';
 
+
 const siteFilterElement = document.querySelector('.trip-controls');
 const siteContentElement = document.querySelector('.trip-events');
+const points = getPoints();
+
 
 const tripPointModel = new TripPointModel();
+tripPointModel.init(points);
 const pointPresenter = new PointPresenter({
   filtersContainer: siteFilterElement,
   mainContainer: siteContentElement,
