+++ src/model/trip-point-model.js
@@ -1,9 +1,11 @@
-import { TRIP_POINT_COUNT, generateTripPoint } from '../mock/point.js';
-
 export default class TripPointModel {
-  #points = Array.from({length: TRIP_POINT_COUNT }, generateTripPoint);
+  #points = [];
 
   get points () {
     return this.#points;
+  }
+
+  init(points){
+    this.#points = points;
   }
 }
