+++ src/utils/common.js
@@ -1,0 +1,10 @@
+const getRandomInteger = (a = 0, b = 1) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+
+  return Math.floor(lower + Math.random() * (upper - lower + 1));
+};
+
+const shuffleElements = (elements) => elements.sort(() => Math.random() - 0.5);
+
+export {getRandomInteger, shuffleElements};
